refactor(pkg-py): replace StateDictAccessorMixin with StateDictQueryChat
Replace the mixin-based approach (which required ~10 type: ignore annotations
for attributes the mixin couldn't own) with a proper subclass of QueryChatBase.
Dash and Gradio QueryChat classes now use single inheritance from
StateDictQueryChat instead of dual inheritance.
